fix: merge conflicts when rebasing
