+++ RailwayCo/Assets/Scripts/GameLogic/Core/Attribute/Attribute.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 05e3e9dceaea9ae44a77481dd32aebac
+guid: ee3f70ad265b74d4c93c7a9cee82972b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
