+++ RailwayCo/Assets/Scripts/GameLogic/Core/Attribute/DoubleAttribute.cs
@@ -1,14 +1,10 @@
 public class DoubleAttribute : Attribute<double>
 {
-    public DoubleAttribute(double lowerLimit, double upperLimit, double amount, double rate)
-    {
-        LowerLimit = lowerLimit;
-        UpperLimit = upperLimit;
-        Amount = amount;
-        Rate = rate;
-    }
+    public DoubleAttribute(double lowerLimit, double upperLimit, double amount, double rate) : base(lowerLimit, upperLimit, amount, rate) { }
 
-    public override void UpgradeLimit(double upgradeAmount) => UpperLimit = DoubleRangeCheck(UpperLimit + upgradeAmount);
+    public override void UpgradeLimit(double upgradeAmount) => UpperLimit = Arithmetic.DoubleRangeCheck(UpperLimit + upgradeAmount);
 
-    public override void UpgradeRate(double upgradeAmount) => Rate = DoubleRangeCheck(Rate + upgradeAmount);
+    public override void UpgradeRate(double upgradeAmount) => Rate = Arithmetic.DoubleRangeCheck(Rate + upgradeAmount);
+    
+    public override object Clone() => new DoubleAttribute(LowerLimit, UpperLimit, Amount, Rate);
 }
